requirements.txt
langchain==0.3.27
langchain-openai==0.3.28
openai==1.98.0
python-gitlab==6.2.0
PyGithub==2.7.0
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
dotenv-python==0.0.1
requests==2.32.4
aiohttp==3.12.15
python-multipart==0.0.20
pydantic-settings==2.6.1